añadir platos al menu

+++ src/main/java/com/example/demo/model/PlatosTable.java
@@ -1,5 +1,7 @@
 package com.example.demo.model;
 
+import java.math.BigDecimal;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
 import org.springframework.data.relational.core.mapping.Table;
@@ -15,7 +17,7 @@ import lombok.NoArgsConstructor;
 public class PlatosTable {
     private @Column("PLATO_ID") @Id Long id;
     private @Column("NOMBRE") String nombre;
-    private @Column("PRECIO") Double precio;
+    private @Column("PRECIO") BigDecimal precio;
     private @Column("FOTO") String foto;
     private @Column("DESCRIPCION") String descripcion;
     private @Column("SECCION") Long seccion;
